feat(security): merge Sandbox + Permissions pages into Security; delete both
User feedback: 'why am i still being forwarded to the sandbox page? we
need to merge them here and then delete, picking custom folders needs to
be included on that main page'.

Custom folders are now fully inline on the Security page inside the
Files expander:
- ListView of granted folders with per-row access combo (Blocked/Read/RW)
  and remove button (Segoe Fluent X glyph).
- 'Add folder…' button uses FolderPicker (initialized via the hub's
  HWND) to add a new grant.
- Selecting 'Blocked' on an existing row removes it.
- Round-trip with settings.SandboxCustomFolders; participates in level
  drift counter.

Pages physically removed from the project:
- src/OpenClaw.Tray.WinUI/Pages/SandboxPage.xaml + .xaml.cs
- src/OpenClaw.Tray.WinUI/Pages/PermissionsPage.xaml + .xaml.cs

HubWindow:
- TagToPageType now maps 'permissions' and 'sandbox' to CapabilitiesPage
  (kept as aliases so any external deep links still land somewhere sane).
- Removed the PermissionsPage/SandboxPage case arms from Initialize.

Capabilities (Security) page now contains every privacy/security control
that used to live on Sandbox or Permissions:
- Sandbox toggle: SystemRunSandboxEnabled (radio: in container / direct)
- Sandbox folder access: Documents/Downloads/Desktop combos
- Sandbox custom folders: inline list with picker  ←  NEW
- Sandbox network: SystemRunAllowOutbound
- Sandbox clipboard: SandboxClipboard radios
- Sandbox limits: timeout + max output
- Exec policy default action + rule editor
